add large number support

+++ computer/IntcodeComputer/AddInstruction.cs
@@ -11,7 +11,7 @@ namespace IntcodeComputer
     public Parameter Parameter2 => Parameters[1];
     public Parameter Parameter3 => Parameters[2];
 
-    public AddInstruction(int opCode, int parameter1, int parameter2, int parameter3)
+    public AddInstruction(int opCode, long parameter1, long parameter2, long parameter3)
     {
       base.OpCode = 1;
       base.PointerAdvancement = 4;
@@ -25,7 +25,7 @@ namespace IntcodeComputer
       CalculateParameterModes(opCode);
     }
 
-    public void Execute(List<int> memory, Dictionary<int, int> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<int> inputBuffer, BufferBlock<int> outputBuffer)
+    public void Execute(List<long> memory, Dictionary<int, long> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<long> inputBuffer, BufferBlock<long> outputBuffer)
     {
       var value1 = GetParameterValue(Parameter1, memory, virtualMemory, relativeBase);
       var value2 = GetParameterValue(Parameter2, memory, virtualMemory, relativeBase);
@@ -34,7 +34,7 @@ namespace IntcodeComputer
       var result = value1 + value2;
 
       // write to memory - this will NEVER be in immediate mode, so it's always an address
-      WriteToMemory(memory, virtualMemory, Parameter3.Value, result);
+      WriteToMemory(memory, virtualMemory, (int)Parameter3.Value, result);
 
       // function point advances 4 places
       instructionPointer += PointerAdvancement;
