+++ computer/IntcodeComputer/IInstruction.cs
@@ -7,6 +7,6 @@ namespace IntcodeComputer
 {
   public interface IInstruction
   {
-    void Execute(List<int> memory, Dictionary<int, int> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<int> inputBuffer, BufferBlock<int> outputBuffer);
+    void Execute(List<long> memory, Dictionary<int, long> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<long> inputBuffer, BufferBlock<long> outputBuffer);
   }
 }
