+++ computer/IntcodeComputer/InputInstruction.cs
@@ -10,7 +10,7 @@ namespace IntcodeComputer
   {
     public Parameter Parameter1 => Parameters[0];
 
-    public InputInstruction(int opCode, int parameter1)
+    public InputInstruction(int opCode, long parameter1)
     {
       base.OpCode = 3;
       base.PointerAdvancement = 2;
@@ -22,18 +22,18 @@ namespace IntcodeComputer
       CalculateParameterModes(opCode);
     }
 
-    public void Execute(List<int> memory, Dictionary<int, int> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<int> inputBuffer, BufferBlock<int> outputBuffer)
+    public void Execute(List<long> memory, Dictionary<int, long> virtualMemory, ref int instructionPointer, ref int relativeBase, BufferBlock<long> inputBuffer, BufferBlock<long> outputBuffer)
     {
       // pull value from input buffer
       var input = ReceiveAsync(inputBuffer).GetAwaiter().GetResult();
 
       // write to memory - this will NEVER be in immediate mode, so it's always an address 
-      WriteToMemory(memory, virtualMemory, Parameter1.Value, input);
+      WriteToMemory(memory, virtualMemory, (int)Parameter1.Value, input);
 
       instructionPointer += PointerAdvancement;
     }
 
-    private async Task<int> ReceiveAsync(BufferBlock<int> inputBuffer)
+    private async Task<long> ReceiveAsync(BufferBlock<long> inputBuffer)
     {
       return await inputBuffer.ReceiveAsync();
     }
