+++ computer/IntcodeComputerTest/ConditionalOppCodeUnitTest.cs
@@ -15,12 +15,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
+      var memory = new List<long>() { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(8)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -36,12 +36,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
+      var memory = new List<long>() { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(9)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -56,12 +56,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
+      var memory = new List<long>() { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(3)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -76,12 +76,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
+      var memory = new List<long>() { 3, 9, 7, 9, 10, 9, 4, 9, 99, -1, 8 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(77)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -96,12 +96,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
+      var memory = new List<long>() { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(8)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -116,12 +116,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
+      var memory = new List<long>() { 3, 3, 1108, -1, 8, 3, 4, 3, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(2)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -136,12 +136,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
+      var memory = new List<long>() { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(2)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -157,12 +157,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
+      var memory = new List<long>() { 3, 3, 1107, -1, 8, 3, 4, 3, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(8)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -177,12 +177,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
+      var memory = new List<long>() { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(0)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -197,12 +197,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
+      var memory = new List<long>() { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(65)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -217,12 +217,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
+      var memory = new List<long>() { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(0)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -237,12 +237,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
+      var memory = new List<long>() { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(765)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -257,12 +257,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99 };
+      var memory = new List<long>() { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(5)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -277,12 +277,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
+      var memory = new List<long>() { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(8)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
@@ -297,12 +297,12 @@ namespace IntcodeComputerTest
     {
       // arrange
       var computer = new Computer();
-      var memory = new List<int>() { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
+      var memory = new List<long>() { 3, 21, 1008, 21, 8, 20, 1005, 20, 22, 107, 8, 21, 20, 1006, 20, 31, 1106, 0, 36, 98, 0, 0, 1002, 21, 125, 20, 4, 20, 1105, 1, 46, 104, 999, 1105, 1, 46, 1101, 1000, 1, 20, 4, 20, 1105, 1, 46, 98, 99 };
 
-      var inputBuffer = new BufferBlock<int>();
+      var inputBuffer = new BufferBlock<long>();
       Task.Run(() => inputBuffer.SendAsync(54378)).Wait();
 
-      var outputBuffer = new BufferBlock<int>();
+      var outputBuffer = new BufferBlock<long>();
 
       // act
       computer.ExecuteProgram(memory, inputBuffer, outputBuffer);
